+++ mate-core/mate-starter-auth/src/main/java/vip/mate/core/auth/annotation/EnableToken.java
@@ -2,6 +2,11 @@ package vip.mate.core.auth.annotation;
 
 import java.lang.annotation.*;
 
+/**
+ * 启用Token验证
+ *
+ * @author pangu
+ */
 @Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
