--- mate-core/mate-starter-security/src/main/java/vip/mate/core/security/config/PermissionService.java
@@ -1,32 +1,0 @@
-package vip.mate.core.security.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.ArrayUtils;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
-import org.springframework.util.PatternMatchUtils;
-import org.springframework.util.StringUtils;
-
-import java.util.Collection;
-
-@Slf4j
-@Service(value = "mp")
-public class PermissionService {
-
-    public boolean hasPerm(String... permissions) {
-        if (ArrayUtils.isEmpty(permissions)) {
-            return false;
-        }
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null) {
-            return false;
-        }
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        return authorities.stream()
-                .map(GrantedAuthority::getAuthority)
-                .filter(StringUtils::hasText)
-                .anyMatch(x -> PatternMatchUtils.simpleMatch(permissions, x));
-    }
-}
