--- mate-core/mate-starter-security/src/main/java/vip/mate/core/security/sms/SmsCodeAuthenticationSecurityConfig.java
@@ -1,47 +1,0 @@
-package vip.mate.core.security.sms;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.stereotype.Component;
-import vip.mate.core.security.userdetails.MateUserDetailsService;
-
-@Component
-public class SmsCodeAuthenticationSecurityConfig
-        extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-    @Autowired
-    @SuppressWarnings("all")
-    private MateUserDetailsService userDetailsService;
-
-    @Autowired
-    @SuppressWarnings("all")
-    public AuthenticationSuccessHandler mateAuthenticationSuccessHandler;
-
-    @Autowired
-    @SuppressWarnings("all")
-    public AuthenticationFailureHandler mateAuthenticationFailureHandler;
-
-    @Override
-    public void configure(HttpSecurity http) {
-
-        // 过滤器
-        SmsCodeAuthenticationFilter smsCodeAuthenticationFilter = new SmsCodeAuthenticationFilter();
-        smsCodeAuthenticationFilter.setAuthenticationManager(http.getSharedObject(AuthenticationManager.class));
-        smsCodeAuthenticationFilter.setAuthenticationSuccessHandler(mateAuthenticationSuccessHandler);
-        smsCodeAuthenticationFilter.setAuthenticationFailureHandler(mateAuthenticationFailureHandler);
-
-
-        // 获取验证码提供者
-        SmsCodeAuthenticationProvider smsCodeAuthenticationProvider = new SmsCodeAuthenticationProvider(userDetailsService);
-
-        // 将短信验证码校验器注册到 HttpSecurity， 并将短信验证码过滤器添加在 UsernamePasswordAuthenticationFilter 之前
-        http.authenticationProvider(smsCodeAuthenticationProvider)
-                .addFilterAfter(smsCodeAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
-    }
-}
