+++ mate-core/mate-starter-security/src/main/resources/META-INF/spring.factories
@@ -1,8 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   vip.mate.core.security.config.IgnoreUrlPropsConfig,\
-  vip.mate.core.security.config.PermissionService,\
-  vip.mate.core.security.social.SocialAuthenticationSecurityConfig,\
-  vip.mate.core.security.sms.SmsCodeAuthenticationSecurityConfig
+  vip.mate.core.security.service.PermissionService
