+++ mate-uaa/src/main/java/vip/mate/uaa/sms/SmsCodeAuthenticationToken.java
@@ -1,4 +1,4 @@
-package vip.mate.core.security.sms;
+package vip.mate.uaa.sms;
 
 import lombok.SneakyThrows;
 import org.springframework.security.authentication.AbstractAuthenticationToken;
